auth_backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.28.2
PyJWT==2.8.0
pymongo==4.6.3
Werkzeug==3.0.2
